// schemas.d.ts
export interface GQLPlan {
  id: string;
  name?: string;
  start?: string;
  end?: string;
  private?: boolean;
  owner?: GQLUser;
  created_at?: string;
  updated_at?: string;
  description?: string;
  counters?: GQLPlanCounters;
  location?: GQLLocation;
  hasLiked?: boolean;
  joinStatus?: GQLJoinType;
  status?: Array<GQLPlanStatus | null>;
  joined?: Array<GQLUser | null>;
  sponsored?: boolean;
  coverImage?: string;
  type?: GQLPlanType;
  activities?: Array<GQLActivity | null>;
}

export interface GQLUser {
  id: string;
  email: string;
  name: string;
  username: string;
  bio?: string;
  visitedCountries?: Array<string | null>;
  phoneNumber?: GQLPhoneNumber;
  created_at?: string;
  image?: string;
  images?: Array<string | null>;
  banned?: boolean;
  location?: GQLLocation;
  friends?: Array<GQLUser | null>;
  mutualFriendsCount?: number;
  friendshipStatus?: GQLFriendshipType;
  notificationRange?: number;
  expoToken?: string;
  tags?: Array<string | null>;
  interests?: Array<GQLActivity | null>;
  home?: GQLHome;
  languages?: Array<string | null>;
  gender?: GQLGender;
  birthDate?: GQLBirthDate;
  website?: string;
}

export interface GQLPhoneNumber {
  countryCode?: string;
  digits?: string;
}

export interface GQLLocation {
  city?: string;
  lat?: number;
  lng?: number;
  country?: string;
  displayName?: string;
  placeId?: string;
  fullAddress?: string;
}

export const enum GQLFriendshipType {
  FRIEND = 'FRIEND',
  REQUESTED = 'REQUESTED',
  REQUESTED_BY = 'REQUESTED_BY',
  BLOCKED = 'BLOCKED',
  BLOCKED_BY = 'BLOCKED_BY',
  STRANGER = 'STRANGER'
}

export const enum GQLActivity {
  HIKING = 'HIKING',
  SCUBADIVING = 'SCUBADIVING',
  SURFING = 'SURFING',
  SKIING = 'SKIING',
  SNOWBOARDING = 'SNOWBOARDING',
  DANCING = 'DANCING',
  SAILING = 'SAILING',
  YOGA = 'YOGA',
  BEACH = 'BEACH',
  RESTAURANTS = 'RESTAURANTS',
  COOKING_CLASSES = 'COOKING_CLASSES',
  LOCAL_MARKETS = 'LOCAL_MARKETS',
  MUSEUMS = 'MUSEUMS',
  HISTORY_TOURS = 'HISTORY_TOURS',
  CULTURAL_FESTIVAL = 'CULTURAL_FESTIVAL',
  WILDLIFE_TOURS = 'WILDLIFE_TOURS',
  EXPLORING_NATURE = 'EXPLORING_NATURE',
  WINE_TASTING = 'WINE_TASTING',
  PARTYING = 'PARTYING',
  CONCERTS = 'CONCERTS',
  LANGUAGE_CLASSES = 'LANGUAGE_CLASSES',
  VOLUNTEERING = 'VOLUNTEERING'
}

export interface GQLHome {
  value?: GQLLocation;
  visibilityLevel?: GQLVisibilityLevel;
}

export const enum GQLVisibilityLevel {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
  FRIENDS_ONLY = 'FRIENDS_ONLY'
}

export interface GQLGender {
  value?: string;
  visibilityLevel?: GQLVisibilityLevel;
}

export interface GQLBirthDate {
  value?: string;
  visibilityLevel?: GQLVisibilityLevel;
}

export interface GQLPlanCounters {
  joined?: number;
  requests?: number;
  invites?: number;
  likes?: number;
  comments?: number;
}

export const enum GQLJoinType {
  UNJOIN = 'UNJOIN',
  REQUESTED = 'REQUESTED',
  INVITED = 'INVITED',
  JOINED = 'JOINED',
  HIDDEN = 'HIDDEN'
}

export const enum GQLPlanStatus {
  OFFICIAL = 'OFFICIAL',
  TENTATIVE = 'TENTATIVE',
  LOOKING_FOR_FRIENDS = 'LOOKING_FOR_FRIENDS',
  LOOKING_FOR_ROOMMATES = 'LOOKING_FOR_ROOMMATES',
  PLANNING = 'PLANNING',
  BOOKED = 'BOOKED'
}

export const enum GQLPlanType {
  RETREAT = 'RETREAT',
  EVENT = 'EVENT'
}

export interface GQLFeedFilter {
  kind?: GQLFeedType;
  country?: string;
}

export const enum GQLFeedType {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE'
}

export interface GQLPageInput {
  count: number;
  page?: number;
  cursor?: string;
}

export interface GQLSort {
  field?: GQLField;
  direction?: GQLDirection;
}

export const enum GQLField {
  UPDATED_AT = 'UPDATED_AT',
  CREATED_AT = 'CREATED_AT',
  START_DATE = 'START_DATE'
}

export const enum GQLDirection {
  ASC = 'ASC',
  DESC = 'DESC'
}

export interface GQLGetPlansResponse {
  results?: Array<GQLPlan | null>;
  page?: GQLPage;
}

export interface GQLPage {
  pageSize?: number;
  hasNext?: boolean;
  total?: number;
  nextCursor?: string;
}
// ...
export interface QueryToUserPlansArgs {
  userId?: string;
  userName?: string;
  desc?: boolean;
  start?: string;
  end?: string;
  page?: GQLPageInput;
  requester?: string;
}